requirements.txt
langchain==0.3.27
langchain-community==0.3.31
langchain-core==0.3.80
langchain-openai==0.2.14
streamlit
openai
tiktoken
chromadb
python-dotenv
duckduckgo-search
pypdf
